# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Service Management API",
        "description": "API documentation for Car Service / Car Wash Management System",
        "contact": {
            "name": "API Support",
            "email": "admin@carservice.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/pricing": {
            "get": {
                "tags": [
                    "Pricing"
                ],
                "summary": "List pricing rules",
                "description": "Get paginated list of vehicle service pricing rules",
                "operationId": "pricingIndex",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "branch_id",
                        "in": "query",
                        "description": "Filter by branch",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "Filter by service",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pricing retrieved successfully"
                    },
                    "400": {
                        "description": "Branch ID is required"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pricing"
                ],
                "summary": "Create pricing rule",
                "description": "Create a new vehicle service pricing rule",
                "operationId": "pricingStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_id",
                                    "service_id",
                                    "vehicle_type_id",
                                    "price"
                                ],
                                "properties": {
                                    "branch_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vehicle_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vehicle_brand_id": {
                                        "description": "Optional for brand-specific pricing",
                                        "type": "integer"
                                    },
                                    "vehicle_model_id": {
                                        "description": "Optional for model-specific pricing",
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 49.99
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pricing created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "409": {
                        "description": "Duplicate pricing rule"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pricing/{id}": {
            "get": {
                "tags": [
                    "Pricing"
                ],
                "summary": "Get pricing rule",
                "description": "Get pricing rule details by ID",
                "operationId": "pricingShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Pricing ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pricing retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Pricing not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pricing"
                ],
                "summary": "Update pricing rule",
                "description": "Update pricing rule details",
                "operationId": "pricingUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Pricing ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_id",
                                    "service_id",
                                    "vehicle_type_id",
                                    "price"
                                ],
                                "properties": {
                                    "branch_id": {
                                        "type": "integer"
                                    },
                                    "service_id": {
                                        "type": "integer"
                                    },
                                    "vehicle_type_id": {
                                        "type": "integer"
                                    },
                                    "vehicle_brand_id": {
                                        "type": "integer"
                                    },
                                    "vehicle_model_id": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pricing updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Pricing not found"
                    },
                    "409": {
                        "description": "Duplicate pricing rule"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pricing"
                ],
                "summary": "Delete pricing rule",
                "description": "Soft delete a pricing rule",
                "operationId": "pricingDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Pricing ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pricing deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Pricing not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pricing/lookup": {
            "get": {
                "tags": [
                    "Pricing"
                ],
                "summary": "Lookup price",
                "description": "Find the best matching price for a service and vehicle combination. Uses hierarchical lookup: exact model -> brand level -> type level",
                "operationId": "pricingLookup",
                "parameters": [
                    {
                        "name": "branch_id",
                        "in": "query",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vehicle_type_id",
                        "in": "query",
                        "description": "Vehicle Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vehicle_brand_id",
                        "in": "query",
                        "description": "Vehicle Brand ID (optional)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vehicle_model_id",
                        "in": "query",
                        "description": "Vehicle Model ID (optional)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Price found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "pricing": {
                                                    "type": "object"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": 49.99
                                                },
                                                "match_type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "exact_model",
                                                        "brand_level",
                                                        "type_level"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "No pricing found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}